Fix iOS narrowing conversion in preview sprite size fallback.
This adds explicit float casts for texture pixel dimensions to satisfy clang++ narrowing rules and keep preview scaling behavior unchanged.

Made-with: Cursor

+++ src/features/capture/ui/CapturePreviewPopup.cpp
@@ -43,7 +43,10 @@ CCSize getSpriteLogicalSize(CCSprite* sprite) {
         return size;
     }
     if (auto* tex = sprite->getTexture()) {
-        return {tex->getPixelsWide(), tex->getPixelsHigh()};
+        return {
+            static_cast<float>(tex->getPixelsWide()),
+            static_cast<float>(tex->getPixelsHigh())
+        };
     }
     return {0.f, 0.f};
 }
